server/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==9.1.0
cryptography==43.0.3


